+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @alosix-hub-ui/react
 
+## 2.1.0
+
+### Minor Changes
+
+- feat: colors contrast
+
 ## 2.0.1
 
 ### Patch Changes